move density_table to new file
