+++ src/PackmolInputCreator/SolutionBoxUSC.jl
@@ -1,25 +1,3 @@
-struct DensityTable
-    concentration::Vector{typeof(1.0u"mol/L")}
-    density::Vector{typeof(1.0u"g/mL")}
-end
-
-function Base.show(io::IO, ::MIME"text/plain", density_table::DensityTable)
-    uc = unit(eltype(density_table.concentration))
-    ud = unit(eltype(density_table.density))
-    print(io, """
-    ==================================================================
-    Density table:
-    ==================================================================
-    Concentration ($uc) |      Density ($ud)
-    """)
-    for (c,d) in zip(density_table.concentration, density_table.density)
-        @printf(io, "%15.3f%9s | %15.3f\n", ustrip(c), ' ', ustrip(d))
-    end
-    print(io, chomp("""
-    ==================================================================
-    """))
-end
-
 mutable struct SolutionBoxUSC <: SolutionBox
     solute_pdbfile::String
     solvent_pdbfile::String
@@ -28,66 +6,6 @@ mutable struct SolutionBoxUSC <: SolutionBox
     solute_molar_mass::Number
     solvent_molar_mass::Number
     cossolvent_molar_mass::Number
-end
-
-function interpolate_density(
-    system,
-    concentration::Number,
-    concentration_units::String,
-)
-    dtab = system.density_table.density
-    ctab = system.density_table.concentration
-    for (ic,c) in enumerate(ctab)
-        cc = cconvert(c, "mol/L" => concentration_units; 
-            M_solvent = system.solvent_molar_mass, 
-            M_solute = system.cossolvent_molar_mass, 
-            rho_solution = dtab[ic],
-            rho_solute = density_pure_cossolvent(system),
-        )
-        if (ic == 1 && concentration < cc) || (ic+1 > length(ctab))
-            cfirst = cconvert(ctab[1], "mol/L" => concentration_units; 
-                M_solvent = system.solvent_molar_mass, 
-                M_solute = system.cossolvent_molar_mass, 
-                rho_solution = dtab[1],
-                rho_solute = density_pure_cossolvent(system),
-            ) 
-            clast = cconvert(ctab[end], "mol/L" => concentration_units; 
-                M_solvent = system.solvent_molar_mass, 
-                M_solute = system.cossolvent_molar_mass, 
-                rho_solution = dtab[end],
-                rho_solute = density_pure_cossolvent(system),
-            )
-            throw(ArgumentError("""\n
-                Concentration out of range of density table.
-                In units of $(concentration_units): $(cfirst) - $(clast)
-                Concentration provided: $(concentration)
-            """))
-        end
-        cnext = cconvert(ctab[ic+1], "mol/L" => concentration_units; 
-            M_solvent = system.solvent_molar_mass, 
-            M_solute = system.cossolvent_molar_mass, 
-            rho_solution = dtab[ic+1],
-            rho_solute = density_pure_cossolvent(system),
-        )
-        if cc <= concentration <= cnext
-            return dtab[ic] + (dtab[ic+1] - dtab[ic]) * (concentration - cc) / (cnext - cc)
-        end
-    end
-end
-
-function DensityTable(
-    concentration_units::String,
-    concentration::AbstractVector{<:Number},
-    density::AbstractVector{<:Number},
-    M_solvent::Quantity,
-    M_cossolvent::Quantity,
-)
-    dvec = typeof(1.0u"g/mL")[uconvert(u"g/mL", 1u"g/mL" * d) for d in density]
-    cvec = typeof(1.0u"mol/L")[
-        cconvert(c, concentration_units => "mol/L"; M_solvent, M_solute=M_cossolvent, rho_solution=d) 
-        for (c,d) in zip(concentration,dvec)
-    ]
-    return DensityTable(cvec, dvec)
 end
 
 """
@@ -155,26 +73,6 @@ function SolutionBoxUSC(;
         cossolvent_molar_mass,
     )
     return system
-end
-
-function density_pure_solvent(system::SolutionBoxUSC) 
-    cfirst = first(system.density_table.concentration)
-    if !(_round(cfirst) ≈ 0.0*cfirst)
-        return missing # pure solvent density is not in the table
-    end
-    first(system.density_table.density)
-end
-function density_pure_cossolvent(system::SolutionBoxUSC) 
-    clast = last(system.density_table.concentration)
-    clast = cconvert(clast, "mol/L" => "x"; 
-        M_solvent = system.solvent_molar_mass, 
-        M_solute = system.cossolvent_molar_mass, 
-        rho_solution = last(system.density_table.density),
-    )
-    if !(_round(clast) ≈ 1.0)
-        return missing # pure cossolvent density is not in the table
-    end
-    last(system.density_table.density[end])
 end
 
 function Base.show(io::IO, ::MIME"text/plain", system::SolutionBoxUSC)
